+++ core/src/pixelcolor/raw/to_bytes.rs
@@ -1,5 +1,5 @@
 use crate::pixelcolor::{
-    raw::{RawU1, RawU16, RawU2, RawU24, RawU32, RawU4, RawU8},
+    raw::{RawU1, RawU12, RawU16, RawU2, RawU24, RawU32, RawU4, RawU8},
     PixelColor,
 };
 
@@ -46,6 +46,39 @@ impl_to_bytes!(RawU4, [u8; 1]);
 impl_to_bytes!(RawU8, [u8; 1]);
 impl_to_bytes!(RawU16, [u8; 2]);
 impl_to_bytes!(RawU32, [u8; 4]);
+
+impl ToBytes for RawU12 {
+    type Bytes = [u8; 2];
+
+    fn to_be_bytes(self) -> Self::Bytes {
+        let mut ret = [0; 2];
+
+        ret.copy_from_slice(&self.0.to_be_bytes());
+
+        ret
+    }
+
+    fn to_le_bytes(self) -> Self::Bytes {
+        let mut ret = [0; 2];
+
+        let le_bytes = &self.0.to_le_bytes();
+
+        ret[0] = le_bytes[0];
+        ret[1] = le_bytes[1] << 4;
+
+        ret
+    }
+
+    #[cfg(target_endian = "big")]
+    fn to_ne_bytes(self) -> Self::Bytes {
+        self.to_be_bytes()
+    }
+
+    #[cfg(target_endian = "little")]
+    fn to_ne_bytes(self) -> Self::Bytes {
+        self.to_le_bytes()
+    }
+}
 
 impl ToBytes for RawU24 {
     type Bytes = [u8; 3];
@@ -116,7 +149,7 @@ where
 mod tests {
     use super::*;
     use crate::pixelcolor::{
-        Bgr565, Bgr666, Bgr888, BinaryColor, Gray2, Gray4, Gray8, Rgb565, Rgb666, Rgb888,
+        Bgr565, Bgr666, Bgr888, BinaryColor, Gray2, Gray4, Gray8, Rgb444, Rgb565, Rgb666, Rgb888,
     };
 
     fn assert_all_orders<T>(value: T, bytes: T::Bytes)
@@ -151,6 +184,38 @@ mod tests {
     fn bpp8() {
         assert_all_orders(Gray8::new(0), [0]);
         assert_all_orders(Gray8::new(255), [255]);
+    }
+
+    #[test]
+    fn bpp12_rgb_be() {
+        assert_eq!(
+            Rgb444::new(15, 0, 0).to_be_bytes(),
+            [0b0000_1111, 0b000_00000]
+        );
+        assert_eq!(
+            Rgb444::new(0, 15, 0).to_be_bytes(),
+            [0b0000_0000, 0b1111_0000]
+        );
+        assert_eq!(
+            Rgb444::new(0, 0, 15).to_be_bytes(),
+            [0b0000_0000, 0b0000_1111]
+        );
+    }
+
+    #[test]
+    fn bpp12_rgb_le() {
+        assert_eq!(
+            Rgb444::new(15, 0, 0).to_le_bytes(),
+            [0b0000_0000, 0b1111_0000]
+        );
+        assert_eq!(
+            Rgb444::new(0, 15, 0).to_le_bytes(),
+            [0b1111_0000, 0b0000_0000]
+        );
+        assert_eq!(
+            Rgb444::new(0, 0, 15).to_le_bytes(),
+            [0b0000_1111, 0b0000_0000]
+        );
     }
 
     #[test]