+++ CHANGELOG.md
@@ -6,9 +6,14 @@ Embedded Graphics is a `no_std` library for adding graphics features to display 
 
 ## [Unreleased] - ReleaseDate
 
+### Changed
+
+- added `Rgb444` and `RawU12` to support 12bit RGB on ST7789 displays.
+
 ### Added
 
 - [#733](https://github.com/embedded-graphics/embedded-graphics/pull/733) Add envelope() method to `Rectangle`
+
 
 ## [0.8.1] - 2023-08-10